+++ common_pyutil/functional.py
@@ -382,10 +382,9 @@ def concat(list_var: Iterable[List]) -> List:
 def takewhile(predicate: Callable[[Any], bool], seq: Iterable) -> Iterable:
     """Lazily evaluated takewhile
 
-    :param predicate: First failure of predicate stops the iteration. Should return bool
-    :param seq: Sequence from which to take
-    :returns: filtered sequence
-    :rtype: Same as `seq`
+    Args:
+        predicate: First failure of predicate stops the iteration. Should return bool
+        seq: Sequence from which to take
 
     """
     it = iter(seq)
@@ -396,6 +395,31 @@ def takewhile(predicate: Callable[[Any], bool], seq: Iterable) -> Iterable:
             _next = it.__next__()
         except StopIteration:
             return None
+
+
+def dropwhile(predicate: Callable[[Any], bool], seq: Iterable) -> Iterable:
+    """Lazily evaluated dropwhile.
+
+    Args:
+        predicate: A `Callable` that returns a bool.
+                   The iterable after first success of predicate is returned.
+        seq: Sequence from which to drop
+
+    """
+    it = iter(seq)
+    _next = it.__next__()
+    while predicate(_next):
+        try:
+            _next = it.__next__()
+        except StopIteration:
+            return None
+    while(_next):
+        yield _next
+        try:
+            _next = it.__next__()
+        except StopIteration:
+            return None
+
 
 
 # FIXME: Broken